+++ mawhrin/comms/__init__.py
@@ -6,4 +6,5 @@ from mawhrin.comms.base import group
 from mawhrin.comms.find import find
 from mawhrin.comms.list import list_
 from mawhrin.comms.make import make
+from mawhrin.comms.repl import repl
 from mawhrin.comms.show import show