+++ tests/test_comms/test_repl.py
@@ -1,0 +1,11 @@
+"""
+Tests for 'mawhrin.comms.repl'.
+"""
+
+
+def test_repl(run):
+    # success - with argument
+    _, rslt = run("repl", "1 + 1")
+    assert rslt.output == "2"
+
+    # cannot test main repl loop